+++ bigtree/tree/modify.py
@@ -1216,6 +1216,10 @@ def copy_or_shift_logic(
 
     # Perform shifting/copying
     for from_path, to_path in zip(from_paths, to_paths):
+        # Reset parameters
+        merge_children2 = merge_children
+        merge_leaves2 = merge_leaves
+
         if with_full_path:
             from_node = search.find_full_path(tree, from_path)
         else:
@@ -1263,7 +1267,7 @@ def copy_or_shift_logic(
                             parent = to_node.parent
                             to_node.parent = None
                             to_node = parent
-                            merge_children = False
+                            merge_children2 = False
                         elif merge_attribute:
                             logging.info(
                                 f"Path {to_path} already exists and their attributes will be merged"
@@ -1278,7 +1282,7 @@ def copy_or_shift_logic(
                             parent = to_node.parent
                             to_node.parent = None
                             to_node = parent
-                            merge_children = False
+                            merge_children2 = False
                         else:
                             logging.info(
                                 f"Path {to_path} already exists and children are merged"
@@ -1303,7 +1307,7 @@ def copy_or_shift_logic(
                             parent = to_node.parent
                             to_node.parent = None
                             to_node = parent
-                            merge_leaves = False
+                            merge_leaves2 = False
                         else:
                             logging.info(
                                 f"Path {to_path} already exists and leaves are merged"
@@ -1346,10 +1350,15 @@ def copy_or_shift_logic(
                     )
 
             # Reassign from_node to new parent
+            # if merge_children and (overriding or merge_attribute):
+            #     merge_children = False
+            # if merge_leaves and merge_attribute:
+            #     merge_leaves = False
             if copy:
                 logging.debug(f"Copying {from_node.node_name}")
                 from_node = from_node.copy()
-            if merge_children:
+            if merge_children2:
+                # overriding / merge_attribute handled merge_children, set merge_children=False
                 logging.debug(
                     f"Reassigning children from {from_node.node_name} to {to_node.node_name}"
                 )
@@ -1358,7 +1367,7 @@ def copy_or_shift_logic(
                         del children.children
                     children.parent = to_node
                 from_node.parent = None
-            elif merge_leaves:
+            elif merge_leaves2:
                 logging.debug(
                     f"Reassigning leaf nodes from {from_node.node_name} to {to_node.node_name}"
                 )